requirements.txt
fastapi
uvicorn 
llama-cpp-python 
pydantic